requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
chromadb==0.4.18
openai==1.3.7
PyMuPDF==1.23.8
python-docx==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy<2.0
httpx<0.25.0 